+++ crates/openwraw/src/raw/mod.rs
@@ -1,6 +1,7 @@
 pub mod chroms;
 pub mod data;
 pub mod extern_inf;
+pub mod func_sts;
 pub mod functions_inf;
 pub mod header;
 pub mod index;